Weapons + Health + Flashlight
Added functionality for weapons. 1 of each of the 5 weapons are randomly
spawned across the maze. The player starts with nothing and must find
one. Press F to toggle the flashlight. I added a script that emulates a
battery on the flashlight; it last for X seconds and takes Y seconds to
recharge. Use it wisely. Players have health and if they die, the scene
resets (test with rocket launcher). ScoreManager contains a point system
that can be used for either zombies killed, or number of keys (or
whatever item) collected.